fix: resolve all 33 clippy warnings in non-test code
- Remove unnecessary casts (u32/u64/u8 -> same type) in pixel.rs, boot.rs, main.rs
- Replace manual Range::contains with (a..b).contains(&x) in bridge.rs, boot.rs, bus.rs, cli.rs
- Replace manual is_multiple_of with .is_multiple_of() in boot.rs
- Use .copied() instead of .map(|s| *s) in cli.rs, main.rs
- Remove useless format!() for static strings in main.rs dashboard
- Remove empty lines after doc comments in canvas.rs
- Replace match with if let for single-pattern destructuring in audio.rs
- Use borrowed expression directly in hermes.rs
- Inline print_literal format args in cli.rs
- Suppress too_many_arguments for handle_terminal_command in canvas.rs
- Add #[allow] for 8-arg function (API stability over refactoring)
- Net: -8 lines, zero clippy warnings on lib+bin

+++ src/pixel.rs
@@ -70,8 +70,8 @@ pub fn decode_rts_png(path: &str) -> Result<DecodedPixels, String> {
         out
     } else {
         // Hilbert curve layout: inverse Hilbert to get linear byte order
-        let grid_w = info.width as u32;
-        let grid_h = info.height as u32;
+        let grid_w = info.width;
+        let grid_h = info.height;
         let grid_side = grid_w.max(grid_h);
         let grid_order = 31 - grid_side.leading_zeros();
